+++ packages/core/src/server/hmr-engine.ts
@@ -8,7 +8,7 @@ import type { Resource } from '@farmfe/runtime/src/resource-loader.js';
 import { UserHmrConfig } from '../config/index.js';
 import type { JsUpdateResult } from '../types/binding.js';
 import { convertErrorMessage } from '../utils/error.js';
-import { bold, formatExecutionTime, green, lightCyan } from '../utils/index.js';
+import { bold, cyan, formatExecutionTime, green } from '../utils/index.js';
 import { WebSocketClient } from './ws.js';
 
 export class HmrEngine {
@@ -83,7 +83,7 @@ export class HmrEngine {
     const result = await this.app.compiler.update(queue);
 
     this.app.logger.info(
-      `${bold(lightCyan(updatedFilesStr))} updated in ${bold(
+      `${bold(cyan(updatedFilesStr))} updated in ${bold(
         green(
           formatExecutionTime(
             performance.now() - start,