+++ js-plugins/vue/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @farmfe/js-plugin-vue
 
+## 4.0.0-nightly-20241023020505
+
+### Patch Changes
+
+- Updated dependencies [3651244]
+  - @farmfe/core@2.0.0-nightly-20241023020505
+
+## 4.0.0-nightly-20241022143924
+
+### Patch Changes
+
+- Updated dependencies [83838bd]
+  - @farmfe/core@2.0.0-nightly-20241022143924
+
+## 4.0.0-nightly-20241022124925
+
+### Patch Changes
+
+- @farmfe/core@2.0.0-nightly-20241022124925
+
 ## 4.0.0-nightly-20241022041556
 
 ### Patch Changes